updated Product Controller, ProductService: some exception handle and validations

+++ src/main/java/com/vbaggio/dscommerce/services/ProductService.java
@@ -2,14 +2,20 @@ package com.vbaggio.dscommerce.services;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.vbaggio.dscommerce.dto.ProductDTO;
 import com.vbaggio.dscommerce.entities.Product;
 import com.vbaggio.dscommerce.repositories.ProductRepository;
+import com.vbaggio.dscommerce.services.exceptions.DatabaseException;
+import com.vbaggio.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,7 +28,7 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException());
 		return modelMapper.map(product, ProductDTO.class);
 	}
 
@@ -41,22 +47,37 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product product = repository.getReferenceById(id);
 
-		modelMapper.getConfiguration().setSkipNullEnabled(true);
-		modelMapper.typeMap(ProductDTO.class, Product.class).addMappings(mapper -> mapper.skip(Product::setId));
+		try {
+			Product product = repository.getReferenceById(id);
 
-		modelMapper.map(dto, product);
+			modelMapper.getConfiguration().setSkipNullEnabled(true);
+			modelMapper.typeMap(ProductDTO.class, Product.class).addMappings(mapper -> mapper.skip(Product::setId));
 
-		modelMapper.getConfiguration().setSkipNullEnabled(false);
+			modelMapper.map(dto, product);
 
-		product = repository.save(product);
+			modelMapper.getConfiguration().setSkipNullEnabled(false);
 
-		return modelMapper.map(product, ProductDTO.class);
+			product = repository.save(product);
+
+			return modelMapper.map(product, ProductDTO.class);
+			
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException();
+		}
+
 	}
-	
-	@Transactional
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id))
+			throw new ResourceNotFoundException();
+		
+		try {
+			repository.deleteById(id);
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Referential integrity violation.");
+		}
 	}
 }
